Day 5 Finished
Shorter day today. Did start sequence text animation and logic, might tweak the animations abit but else it works. Timings are now off on trail genrations because they dont start at gamestate changes. I have started on player abilitys. Missing Rounds logic for the full game loop. Also needs to add extra key inputs for player 2-4. Getting closer.

+++ Assets/Scripts/UI/TogglePlayers.cs
@@ -22,7 +22,7 @@ public class TogglePlayers : MonoBehaviour
     public void OnClick()
     {   //TODO save the addition or subtraction to a gamestate script aswell for initialization of game. 
 
-        float _currentAmount = float.Parse(_numberOfPlayers.text);  //saving as float for math.
+        int _currentAmount = int.Parse(_numberOfPlayers.text);  //saving as float for math.
         if (_additive)
         {
             //need to check size of current players.
